+++ plugins/short_link/data/sql/schema.sql
@@ -12,8 +12,8 @@ DROP TABLE IF EXISTS `short_link`;
 
 CREATE TABLE `short_link`
 (
-	`id` VARCHAR(5)  NOT NULL,
-	`int_id` INTEGER  NOT NULL AUTO_INCREMENT,
+	`id` VARCHAR(20)  NOT NULL,
+	`int_id` BIGINT NOT NULL AUTO_INCREMENT,
 	`created_at` DATETIME,
 	`updated_at` DATETIME,
 	`expires_at` DATETIME,
@@ -23,6 +23,7 @@ CREATE TABLE `short_link`
 	`system_name` VARCHAR(63),
 	`full_url` VARCHAR(255),
 	`status` INTEGER,
+	`custom_data` text,
 	PRIMARY KEY (`id`),
 	KEY `int_id`(`int_id`),
 	KEY `partner_id`(`partner_id`),