+++ plugins/short_link/lib/model/ShortLink.php
@@ -100,7 +100,7 @@ class ShortLink extends BaseShortLink implements IBaseObject {
 			if(!$dcChar)
 				$dcChar = '0';
 				
-			$id = $dcChar . kString::generateStringId(4);
+			$id = $dcChar . kString::generateStringId(7);
 			ShortLinkPeer::setUseCriteriaFilter(false);
 			$existingObject = ShortLinkPeer::retrieveByPK($id);
 			ShortLinkPeer::setUseCriteriaFilter(true);