+++ plugins/short_link/lib/model/om/BaseShortLink.php
@@ -27,7 +27,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 
 	/**
 	 * The value for the int_id field.
-	 * @var        int
+	 * @var        string
 	 */
 	protected $int_id;
 
@@ -84,6 +84,12 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 	 * @var        int
 	 */
 	protected $status;
+
+	/**
+	 * The value for the custom_data field.
+	 * @var        string
+	 */
+	protected $custom_data;
 
 	/**
 	 * Flag to prevent endless save loop, if this object is referenced
@@ -143,7 +149,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 	/**
 	 * Get the [int_id] column value.
 	 * 
-	 * @return     int
+	 * @return     string
 	 */
 	public function getIntId()
 	{
@@ -331,6 +337,16 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 	}
 
 	/**
+	 * Get the [custom_data] column value.
+	 * 
+	 * @return     string
+	 */
+	public function getCustomData()
+	{
+		return $this->custom_data;
+	}
+
+	/**
 	 * Set the value of [id] column.
 	 * 
 	 * @param      string $v new value
@@ -356,7 +372,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 	/**
 	 * Set the value of [int_id] column.
 	 * 
-	 * @param      int $v new value
+	 * @param      string $v new value
 	 * @return     ShortLink The current object (for fluent API support)
 	 */
 	public function setIntId($v)
@@ -365,7 +381,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 			$this->oldColumnsValues[ShortLinkPeer::INT_ID] = $this->int_id;
 
 		if ($v !== null) {
-			$v = (int) $v;
+			$v = (string) $v;
 		}
 
 		if ($this->int_id !== $v) {
@@ -665,6 +681,26 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 	} // setStatus()
 
 	/**
+	 * Set the value of [custom_data] column.
+	 * 
+	 * @param      string $v new value
+	 * @return     ShortLink The current object (for fluent API support)
+	 */
+	public function setCustomData($v)
+	{
+		if ($v !== null) {
+			$v = (string) $v;
+		}
+
+		if ($this->custom_data !== $v) {
+			$this->custom_data = $v;
+			$this->modifiedColumns[] = ShortLinkPeer::CUSTOM_DATA;
+		}
+
+		return $this;
+	} // setCustomData()
+
+	/**
 	 * Indicates whether the columns in this object are only set to default values.
 	 *
 	 * This method can be used in conjunction with isModified() to indicate whether an object is both
@@ -695,11 +731,14 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 	public function hydrate($row, $startcol = 0, $rehydrate = false)
 	{
 		$this->last_hydrate_time = time();
-
+		
+		// Nullify cached objects
+		$this->m_custom_data = null;
+		
 		try {
 
 			$this->id = ($row[$startcol + 0] !== null) ? (string) $row[$startcol + 0] : null;
-			$this->int_id = ($row[$startcol + 1] !== null) ? (int) $row[$startcol + 1] : null;
+			$this->int_id = ($row[$startcol + 1] !== null) ? (string) $row[$startcol + 1] : null;
 			$this->created_at = ($row[$startcol + 2] !== null) ? (string) $row[$startcol + 2] : null;
 			$this->updated_at = ($row[$startcol + 3] !== null) ? (string) $row[$startcol + 3] : null;
 			$this->expires_at = ($row[$startcol + 4] !== null) ? (string) $row[$startcol + 4] : null;
@@ -709,6 +748,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 			$this->system_name = ($row[$startcol + 8] !== null) ? (string) $row[$startcol + 8] : null;
 			$this->full_url = ($row[$startcol + 9] !== null) ? (string) $row[$startcol + 9] : null;
 			$this->status = ($row[$startcol + 10] !== null) ? (int) $row[$startcol + 10] : null;
+			$this->custom_data = ($row[$startcol + 11] !== null) ? (string) $row[$startcol + 11] : null;
 			$this->resetModified();
 
 			$this->setNew(false);
@@ -718,7 +758,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 			}
 
 			// FIXME - using NUM_COLUMNS may be clearer.
-			return $startcol + 11; // 11 = ShortLinkPeer::NUM_COLUMNS - ShortLinkPeer::NUM_LAZY_LOAD_COLUMNS).
+			return $startcol + 12; // 12 = ShortLinkPeer::NUM_COLUMNS - ShortLinkPeer::NUM_LAZY_LOAD_COLUMNS).
 
 		} catch (Exception $e) {
 			throw new PropelException("Error populating ShortLink object", $e);
@@ -771,7 +811,9 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 		// already in the pool.
 
 		ShortLinkPeer::setUseCriteriaFilter(false);
-		$stmt = ShortLinkPeer::doSelectStmt($this->buildPkeyCriteria(), $con);
+		$criteria = $this->buildPkeyCriteria();
+		ShortLinkPeer::addSelectColumns($criteria);
+		$stmt = BasePeer::doSelect($criteria, $con);
 		ShortLinkPeer::setUseCriteriaFilter(true);
 		$row = $stmt->fetch(PDO::FETCH_NUM);
 		$stmt->closeCursor();
@@ -858,18 +900,84 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 			} else {
 				$ret = $ret && $this->preUpdate($con);
 			}
-			if ($ret) {
-				$affectedRows = $this->doSave($con, $skipReload);
-				if ($isInsert) {
-					$this->postInsert($con);
-				} else {
-					$this->postUpdate($con);
-				}
-				$this->postSave($con);
-				ShortLinkPeer::addInstanceToPool($this);
-			} else {
-				$affectedRows = 0;
+			
+			if (!$ret || !$this->isModified()) {
+				$con->commit();
+				return 0;
 			}
+			
+			for ($retries = 1; $retries < KalturaPDO::SAVE_MAX_RETRIES; $retries++)
+			{
+               $affectedRows = $this->doSave($con);
+                if ($affectedRows || !$this->isColumnModified(ShortLinkPeer::CUSTOM_DATA)) //ask if custom_data wasn't modified to avoid retry with atomic column 
+                	break;
+
+                KalturaLog::debug("was unable to save! retrying for the $retries time");
+                $criteria = $this->buildPkeyCriteria();
+				$criteria->addSelectColumn(ShortLinkPeer::CUSTOM_DATA);
+                $stmt = BasePeer::doSelect($criteria, $con);
+                $cutsomDataArr = $stmt->fetchAll(PDO::FETCH_COLUMN);
+                $newCustomData = $cutsomDataArr[0];
+
+                $this->custom_data_md5 = is_null($newCustomData) ? null : md5($newCustomData);
+
+                $valuesToChangeTo = $this->m_custom_data->toArray();
+				$this->m_custom_data = myCustomData::fromString($newCustomData); 
+
+				//set custom data column values we wanted to change to
+				$validUpdate = true;
+				$atomicCustomDataFields = ShortLinkPeer::getAtomicCustomDataFields();
+			 	foreach ($this->oldCustomDataValues as $namespace => $namespaceValues){
+                	foreach($namespaceValues as $name => $oldValue)
+					{
+						$atomicField = false;
+						if($namespace) {
+							$atomicField = array_key_exists($namespace, $atomicCustomDataFields) && in_array($name, $atomicCustomDataFields[$namespace]);
+						} else {
+							$atomicField = in_array($name, $atomicCustomDataFields);
+						}
+						if($atomicField) {
+							$dbValue = $this->m_custom_data->get($name, $namespace);
+							if($oldValue != $dbValue) {
+								$validUpdate = false;
+								break;
+							}
+						}
+						
+						$newValue = null;
+						if ($namespace)
+						{
+							if (isset ($valuesToChangeTo[$namespace][$name]))
+								$newValue = $valuesToChangeTo[$namespace][$name];
+						}
+						else
+						{ 
+							$newValue = $valuesToChangeTo[$name];
+						}
+		
+						if (is_null($newValue)) {
+							$this->removeFromCustomData($name, $namespace);
+						}
+						else {
+							$this->putInCustomData($name, $newValue, $namespace);
+						}
+					}
+				}
+                   
+				if(!$validUpdate) 
+					break;
+					                   
+				$this->setCustomData($this->m_custom_data->toString());
+			}
+
+			if ($isInsert) {
+				$this->postInsert($con);
+			} else {
+				$this->postUpdate($con);
+			}
+			$this->postSave($con);
+			ShortLinkPeer::addInstanceToPool($this);
+			
 			$con->commit();
 			return $affectedRows;
 		} catch (PropelException $e) {
@@ -956,10 +1064,12 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 	/**
 	 * Code to be run before persisting the object
 	 * @param PropelPDO $con
-	 * @return bloolean
+	 * @return boolean
 	 */
 	public function preSave(PropelPDO $con = null)
 	{
+		$this->setCustomDataObj();
+    	
 		return parent::preSave($con);
 	}
 
@@ -970,7 +1080,9 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 	public function postSave(PropelPDO $con = null) 
 	{
 		kEventsManager::raiseEvent(new kObjectSavedEvent($this));
-		$this->oldColumnsValues = array(); 
+		$this->oldColumnsValues = array();
+		$this->oldCustomDataValues = array();
+    	 
 		parent::postSave($con);
 	}
 	
@@ -981,8 +1093,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 	 */
 	public function preInsert(PropelPDO $con = null)
 	{
-    	$this->setCreatedAt(time());
-    	
+		$this->setCreatedAt(time());
 		$this->setUpdatedAt(time());
 		return parent::preInsert($con);
 	}
@@ -1017,7 +1128,9 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 		if($this->isModified())
 		{
 			kQueryCache::invalidateQueryCache($this);
-			kEventsManager::raiseEvent(new kObjectChangedEvent($this, $this->tempModifiedColumns));
+			$modifiedColumns = $this->tempModifiedColumns;
+			$modifiedColumns[kObjectChangedEvent::CUSTOM_DATA_OLD_VALUES] = $this->oldCustomDataValues;
+			kEventsManager::raiseEvent(new kObjectChangedEvent($this, $modifiedColumns));
 		}
 			
 		$this->tempModifiedColumns = array();
@@ -1208,6 +1321,9 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 			case 10:
 				return $this->getStatus();
 				break;
+			case 11:
+				return $this->getCustomData();
+				break;
 			default:
 				return null;
 				break;
@@ -1240,6 +1356,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 			$keys[8] => $this->getSystemName(),
 			$keys[9] => $this->getFullUrl(),
 			$keys[10] => $this->getStatus(),
+			$keys[11] => $this->getCustomData(),
 		);
 		return $result;
 	}
@@ -1304,6 +1421,9 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 			case 10:
 				$this->setStatus($value);
 				break;
+			case 11:
+				$this->setCustomData($value);
+				break;
 		} // switch()
 	}
 
@@ -1339,6 +1459,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 		if (array_key_exists($keys[8], $arr)) $this->setSystemName($arr[$keys[8]]);
 		if (array_key_exists($keys[9], $arr)) $this->setFullUrl($arr[$keys[9]]);
 		if (array_key_exists($keys[10], $arr)) $this->setStatus($arr[$keys[10]]);
+		if (array_key_exists($keys[11], $arr)) $this->setCustomData($arr[$keys[11]]);
 	}
 
 	/**
@@ -1361,6 +1482,7 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 		if ($this->isColumnModified(ShortLinkPeer::SYSTEM_NAME)) $criteria->add(ShortLinkPeer::SYSTEM_NAME, $this->system_name);
 		if ($this->isColumnModified(ShortLinkPeer::FULL_URL)) $criteria->add(ShortLinkPeer::FULL_URL, $this->full_url);
 		if ($this->isColumnModified(ShortLinkPeer::STATUS)) $criteria->add(ShortLinkPeer::STATUS, $this->status);
+		if ($this->isColumnModified(ShortLinkPeer::CUSTOM_DATA)) $criteria->add(ShortLinkPeer::CUSTOM_DATA, $this->custom_data);
 
 		return $criteria;
 	}
@@ -1379,17 +1501,37 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 
 		$criteria->add(ShortLinkPeer::ID, $this->id);
 		
-		if($this->alreadyInSave && count($this->modifiedColumns) == 2 && $this->isColumnModified(ShortLinkPeer::UPDATED_AT))
+		if($this->alreadyInSave)
 		{
-			$theModifiedColumn = null;
-			foreach($this->modifiedColumns as $modifiedColumn)
-				if($modifiedColumn != ShortLinkPeer::UPDATED_AT)
-					$theModifiedColumn = $modifiedColumn;
-					
-			$atomicColumns = ShortLinkPeer::getAtomicColumns();
-			if(in_array($theModifiedColumn, $atomicColumns))
-				$criteria->add($theModifiedColumn, $this->getByName($theModifiedColumn, BasePeer::TYPE_COLNAME), Criteria::NOT_EQUAL);
-		}
+			if ($this->isColumnModified(ShortLinkPeer::CUSTOM_DATA))
+			{
+				if (!is_null($this->custom_data_md5))
+				{
+					$criteria->add(ShortLinkPeer::CUSTOM_DATA, "MD5(cast(" . ShortLinkPeer::CUSTOM_DATA . " as char character set latin1)) = '$this->custom_data_md5'", Criteria::CUSTOM);
+					//casting to latin char set to avoid mysql and php md5 difference
+					if (kDataCenterMgr::isMultiDc()) // if multi DC configuration don't check costume data on other DC
+					{
+						$currentDcId = kDataCenterMgr::getCurrentDcId();
+						//addOr(column, value, comparison)
+						$criteria->addOr(ShortLinkPeer::CUSTOM_DATA," '$currentDcId' != getDC()" ,Criteria::CUSTOM);
+					}
+				}
+				else 
+					$criteria->add(ShortLinkPeer::CUSTOM_DATA, NULL, Criteria::ISNULL);
+			}
+			
+			if (count($this->modifiedColumns) == 2 && $this->isColumnModified(ShortLinkPeer::UPDATED_AT))
+			{
+				$theModifiedColumn = null;
+				foreach($this->modifiedColumns as $modifiedColumn)
+					if($modifiedColumn != ShortLinkPeer::UPDATED_AT)
+						$theModifiedColumn = $modifiedColumn;
+						
+				$atomicColumns = ShortLinkPeer::getAtomicColumns();
+				if(in_array($theModifiedColumn, $atomicColumns))
+					$criteria->add($theModifiedColumn, $this->getByName($theModifiedColumn, BasePeer::TYPE_COLNAME), Criteria::NOT_EQUAL);
+			}
+		}		
 
 		return $criteria;
 	}
@@ -1446,6 +1588,8 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 		$copyObj->setFullUrl($this->full_url);
 
 		$copyObj->setStatus($this->status);
+
+		$copyObj->setCustomData($this->custom_data);
 
 
 		$copyObj->setNew(true);
@@ -1526,6 +1670,153 @@ abstract class BaseShortLink extends BaseObject  implements Persistent {
 
 	}
 
+	/* ---------------------- CustomData functions ------------------------- */
+
+	/**
+	 * @var myCustomData
+	 */
+	protected $m_custom_data = null;
+	
+	/**
+	 * The md5 value for the custom_data field.
+	 * @var        string
+	 */
+	protected $custom_data_md5;
+
+	/**
+	 * Store custom data old values before the changes
+	 * @var        array
+	 */
+	protected $oldCustomDataValues = array();
+	
+	/**
+	 * @return array
+	 */
+	public function getCustomDataOldValues()
+	{
+		return $this->oldCustomDataValues;
+	}
+	
+	/**
+	 * @param string $name
+	 * @param string $value
+	 * @param string $namespace
+	 * @return string
+	 */
+	public function putInCustomData ( $name , $value , $namespace = null )
+	{
+		$customData = $this->getCustomDataObj( );
+		
+		$customDataOldValue = $customData->get($name, $namespace);
+		if(!is_null($customDataOldValue) && serialize($customDataOldValue) === serialize($value))
+			return;
+				
+		$currentNamespace = '';
+		if($namespace)
+			$currentNamespace = $namespace;
+			
+		if(!isset($this->oldCustomDataValues[$currentNamespace]))
+			$this->oldCustomDataValues[$currentNamespace] = array();
+		if(!isset($this->oldCustomDataValues[$currentNamespace][$name]))
+			$this->oldCustomDataValues[$currentNamespace][$name] = $customDataOldValue;
+		
+		$customData->put ( $name , $value , $namespace );
+	}
+
+	/**
+	 * @param string $name
+	 * @param string $namespace
+	 * @param string $defaultValue
+	 * @return string
+	 */
+	public function getFromCustomData ( $name , $namespace = null , $defaultValue = null )
+	{
+		$customData = $this->getCustomDataObj( );
+		$res = $customData->get ( $name , $namespace );
+		if ( $res === null ) return $defaultValue;
+		return $res;
+	}
+
+	/**
+	 * @param string $name
+	 * @param string $namespace
+	 */
+	public function removeFromCustomData ( $name , $namespace = null)
+	{
+		$customData = $this->getCustomDataObj();
+		
+		$currentNamespace = '';
+		if($namespace)
+			$currentNamespace = $namespace;
+			
+		if(!isset($this->oldCustomDataValues[$currentNamespace]))
+			$this->oldCustomDataValues[$currentNamespace] = array();
+		if(!isset($this->oldCustomDataValues[$currentNamespace][$name]))
+			$this->oldCustomDataValues[$currentNamespace][$name] = $customData->get($name, $namespace);
+		
+		return $customData->remove ( $name , $namespace );
+	}
+
+	/**
+	 * @param string $name
+	 * @param int $delta
+	 * @param string $namespace
+	 * @return string
+	 */
+	public function incInCustomData ( $name , $delta = 1, $namespace = null)
+	{
+		$customData = $this->getCustomDataObj( );
+		
+		$currentNamespace = '';
+		if($namespace)
+			$currentNamespace = $namespace;
+			
+		if(!isset($this->oldCustomDataValues[$currentNamespace]))
+			$this->oldCustomDataValues[$currentNamespace] = array();
+		if(!isset($this->oldCustomDataValues[$currentNamespace][$name]))
+			$this->oldCustomDataValues[$currentNamespace][$name] = $customData->get($name, $namespace);
+		
+		return $customData->inc ( $name , $delta , $namespace  );
+	}
+
+	/**
+	 * @param string $name
+	 * @param int $delta
+	 * @param string $namespace
+	 * @return string
+	 */
+	public function decInCustomData ( $name , $delta = 1, $namespace = null)
+	{
+		$customData = $this->getCustomDataObj(  );
+		return $customData->dec ( $name , $delta , $namespace );
+	}
+
+	/**
+	 * @return myCustomData
+	 */
+	public function getCustomDataObj( )
+	{
+		if ( ! $this->m_custom_data )
+		{
+			$this->m_custom_data = myCustomData::fromString ( $this->getCustomData() );
+		}
+		return $this->m_custom_data;
+	}
+	
+	/**
+	 * Must be called before saving the object
+	 */
+	public function setCustomDataObj()
+	{
+		if ( $this->m_custom_data != null )
+		{
+			$this->custom_data_md5 = is_null($this->custom_data) ? null : md5($this->custom_data);
+			$this->setCustomData( $this->m_custom_data->toString() );
+		}
+	}
+	
+	/* ---------------------- CustomData functions ------------------------- */
+	
 	protected $last_hydrate_time;
 
 	public function getLastHydrateTime()