+++ release-notes.md
@@ -1,3 +1,12 @@
+# Venus-22.13.0
+## Extend short link to support larger volume of objects ##
+- Issue Type: Task
+- Issue ID: PLAT-25698
+
+### Deployment Scripts ###
+
+    mysql –h{HOSTNAME} –u{USER} –p{PASSWORD} kaltura < deployment/updates/sql/2023_07_20_alter_short_link_table.sql
+
 # Venus-22.12.0
 ## Add 'dash' tag to  audio description flavor params ##
 - Issue Type: Task